+++ pytest_indexd/fixtures/indexd_settings.py
