+++ pytest_indexd/fixtures/loader_fixture.py
@@ -1,26 +1,12 @@
-from typing import Iterable, List, Tuple, Union
+from typing import Iterable, List
 
 import pytest
 import yaml
-from _pytest.fixtures import fixture
 
 from indexclient.client import Document, IndexClient
 from indexclient.types import IndexData
 from pytest_indexd import hints
-from pytest_indexd.indexd_server_fixture import MockServer
 from pytest_indexd.utils import mock_doc
-
-
-@pytest.fixture
-def indexd_client(
-    indexd_server: MockServer,
-    create_indexd_tables: fixture,  # usefixtures will cause failure for this one
-    indexd_admin_user: Tuple[str, str],
-) -> IndexClient:
-    """Create the tables and add an auth user"""
-    return IndexClient(
-        indexd_server.baseurl, auth=(indexd_admin_user[0], indexd_admin_user[1])
-    )
 
 
 @pytest.fixture()
