+++ pytest_indexd/plugin.py
@@ -7,5 +7,6 @@ from pytest_indexd.db import (
     pg_url,
     postgresql_server_indexd,
 )
-from pytest_indexd.indexd_client_fixture import indexd_client, indexd_loader
-from pytest_indexd.indexd_server_fixture import indexd_server
+from pytest_indexd.fixtures.client_fixture import indexd_client
+from pytest_indexd.fixtures.loader_fixture import indexd_loader
+from pytest_indexd.fixtures.server_fixture import indexd_server
